Real fix for the gnome "crash."   What was happening was when there was no input device, no midi_reader process was created, and the pid I stored for it was -1.  Well, when you send a signal to process -1, that means send it to all processes with pids > 0.  E.g. to everything, including the window manager.  So, on exit, I was sending a signal 15 (terminate) to not just the gneutronica processes, but to EVERY process...   D'oh!

+++ gneutronica.c
@@ -2817,8 +2817,10 @@ void destroy_event(GtkWidget *widget,
 
 	printf("\n\n\n\n\n\n      Gneutronica quits... your File is saved in /tmp/zzz.gdt\n\n\n\n\n\n");
 	save_to_file("/tmp/zzz.gdt");
-	kill(player_process_pid, SIGTERM); /* a little brutal, but effective . . . */
-	kill(midi_reader_process_id, SIGTERM);
+	if (player_process_pid != -1)
+		kill(player_process_pid, SIGTERM); /* a little brutal, but effective . . . */
+	if (midi_reader_process_id != -1)
+		kill(midi_reader_process_id, SIGTERM);
 	gtk_main_quit();
 }
